+++ src/App.jsx
@@ -1,23 +1,32 @@
-import { useState } from "react";
-
+/* ====================STYLES==================== */
 import "./App.css";
 import "@assets/resources/styles/css/style.css";
+
+/* ====================REACT==================== */
+import { useState } from "react";
+
+/* ====================R-R-D==================== */
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+
+/* ====================COMPONENTS==================== */
 import { Header } from "./main/dev_abhi/components/layout/header/Header";
-import { Home } from "./main/dev_abhi/pages/home/Home";
 import { Footer } from "./main/dev_abhi/components/layout/footer/Footer";
-import { Route, Routes } from "react-router-dom";
+
+/* ====================PAGES==================== */
+import { Home } from "./main/dev_abhi/pages/home/Home";
 import { NotFound } from "./main/dev_abhi/pages/notFound/NotFound";
+/* ============================================================ */
 
 function App() {
 	return (
-		<>
+		<BrowserRouter>
 			<Header />
 			<Routes>
 				<Route path="/" element={<Home />} />
 				<Route path="/*" element={<NotFound />} />
 			</Routes>
 			<Footer />
-		</>
+		</BrowserRouter>
 	);
 }