+++ src/main/dev_abhi/components/ImgTag.jsx
@@ -1,0 +1,5 @@
+const ImgTag = function ({ imgUrl }) {
+	return <img src={imgUrl} draggable className="img" crossOrigin="anonymous" />;
+};
+
+export { ImgTag };