+++ src/main/dev_abhi/pages/home/Home.jsx
@@ -1,12 +1,18 @@
+/* ====================REACT==================== */
+import { useRef } from "react";
+
+/* ====================COMPONENTS==================== */
 import { MainImgContainer } from "../../components/MainImgContainer";
 import { PhotoOperationsContainer } from "../../components/PhotoOperationsContainer";
+/* ============================================================ */
 
 const Home = function () {
+	const downloadContext = useRef(null);
 	return (
-		<>
-			<MainImgContainer />
-			<PhotoOperationsContainer />
-		</>
+		<section className="sectionHome">
+			<MainImgContainer downloadContext={downloadContext} />
+			<PhotoOperationsContainer downloadContext={downloadContext} />
+		</section>
 	);
 };