+++ src/main/dev_abhi/shared/constants.jsx
@@ -5,6 +5,11 @@
 // import { icons, images } from "@assets/resources/js/pathShare";
 import assets from "@assets/resources/js/pathShare";
 
+const DRAGOVER = "dragOver";
+const DROP = "drop";
+const DRAGLEAVE = "dragleave";
+
 // export default images;
 // export { icons };
 export default assets;
+export { DRAGOVER, DROP, DRAGLEAVE };