+++ src/main/dev_abhi/utils/ImgTag.jsx
@@ -1,5 +1,0 @@
-const ImgTag = function ({ imgUrl }) {
-	return <img src={imgUrl} className="img" crossOrigin="anonymous" />;
-};
-
-export { ImgTag };